Refactor, scroll on widget toggle
